backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.2.0"
description = "Phase II: FastAPI Backend with Better Auth JWT"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Tasmeer Jamali" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
    "httpx>=0.28.0",
    # Phase III: AI Chatbot (Gemini for generous free tier)
    "google-generativeai>=0.8.0",
]

[project.scripts]
backend = "src.main:run_server"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src"]